+++ teneto/neuroimagingtools/__init__.py
@@ -4,9 +4,10 @@ from .fmriutils import make_parcellation
 from .bidsutils import make_directories, drop_bids_suffix, \
     get_bids_tag, load_tabular_file, \
     get_sidecar, confound_matching, \
-    process_exclusion_criteria
+    process_exclusion_criteria, \
+    censor_timepoints, exclude_runs
 
-__all__ = ['make_parcellation', 'tnet_to_nx', 'make_directories', 'drop_bids_suffix',
+__all__ = ['make_parcellation', 'make_directories', 'drop_bids_suffix',
            'get_bids_tag', 'load_tabular_file',
            'get_sidecar', 'confound_matching',
-           'process_exclusion_criteria']
+           'process_exclusion_criteria', 'exclude_runs', 'censor_timepoints']